+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.7] - 2026-06-06
+
+### New features
+- Replaced the basic in-app update checker with the industry-standard Sparkle 2 framework.
+- Added secure, automatic in-app updates using EdDSA cryptographic signatures.
+
 ## [0.5.6] - 2026-06-05
 
 ### New features